requirements.txt
flask==3.0.0
flask-cors==4.0.0
anthropic==0.40.0
PyPDF2==3.0.1
Pillow==10.2.0
pytesseract==0.3.10
